Set priorityClass for Kube Proxy
Also, set equal CPU and Memory requests and limits so that Kube Proxy will get Guaranteed QoS. This is required per https://v1-9.docs.kubernetes.io/docs/tasks/administer-cluster/out-of-resource/ ("If necessary, kubelet evicts Pods one at a time to reclaim disk when DiskPressure is encountered. If the kubelet is responding to inode starvation, it reclaims inodes by evicting Pods with the lowest quality of service first. If the kubelet is responding to lack of available disk, it ranks Pods within a quality of service that consumes the largest amount of disk and kill those first.")